Change: Update Version Number to 2.1.7. Change: Add sorting groups for file types into the Qbs project file.
Fix: Remove many implicit conversions from quint64, quint32, uint, int to other various similar types. In other cases we wrap these implicit conversions with a static_cast of the appropriate target type.

+++ src/comments.cpp
@@ -44,8 +44,8 @@ void Comments::newUserCommentSlot(QString& sernum, QString& alias,
     if ( obj == nullptr )
         return;
 
-    quint64 date = QDateTime::currentDateTime()
-                        .toTime_t();
+    uint date = QDateTime::currentDateTime()
+                     .toTime_t();
     QString comment = QString( "\r\n --- \r\n"
                                "%1 \r\n"
                                "SerNum: %2 \r\n"
